pyproject.toml
[project]
name = "jobber"
version = "0.0.3"
description = "Minimal build/push/submit helpers for Docker, AWS S3/ECR/SageMaker, and GCP GCS/GCR/Vertex AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "sagemaker",
    "pyyaml",
    "google-cloud-aiplatform",
]
optional-dependencies.dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["jobber"]
include = ["jobber/templates/*.Dockerfile"]

[project.scripts]
jobber = "jobber.cli:main"
